requirements.txt
torch>=2.4.0
torchvision>=0.19.0
pillow>=10.0.0
numpy>=1.26.0
accelerate>=0.30.0
